# [SPARK-39489][CORE] Improve event logging JsonProtocol performance by using Jackson instead of Json4s