--- a/core/src/main/scala/org/apache/spark/deploy/history/EventFilter.scala
+++ b/core/src/main/scala/org/apache/spark/deploy/history/EventFilter.scala
@@ -59,7 +59,6 @@ private[spark] trait EventFilter {
 }
 
 private[spark] object EventFilter extends Logging {
-
   case class FilterStatistics(
       totalJobs: Long,
       liveJobs: Long,