--- a/core/src/main/scala/org/apache/spark/resource/ResourceInformation.scala
+++ b/core/src/main/scala/org/apache/spark/resource/ResourceInformation.scala
@@ -55,6 +55,8 @@ class ResourceInformation(
 
   override def hashCode(): Int = Seq(name, addresses.toSeq).hashCode()
 
+  // TODO(SPARK-39658): reconsider whether we want to expose a third-party library's
+  // symbols as part of a public API:
   final def toJson(): JValue = ResourceInformationJson(name, addresses).toJValue
 }