--- a/core/src/main/scala/org/apache/spark/util/JsonProtocol.scala
+++ b/core/src/main/scala/org/apache/spark/util/JsonProtocol.scala
@@ -27,6 +27,7 @@ import scala.collection.Map
 import com.fasterxml.jackson.core.{JsonEncoding, JsonGenerator}
 import com.fasterxml.jackson.databind.{DeserializationFeature, JsonNode, ObjectMapper}
 import com.fasterxml.jackson.module.scala.DefaultScalaModule
+import org.json4s.jackson.JsonMethods.compact
 
 import org.apache.spark._
 import org.apache.spark.executor._
@@ -689,7 +690,7 @@ private[spark] object JsonProtocol {
     g.writeObjectFieldStart("Resources")
     executorInfo.resourcesInfo.foreach { case (k, v) =>
       g.writeFieldName(k)
-      v.writeJson(g)
+      g.writeRawValue(compact(v.toJson()))
     }
     g.writeEndObject()
     g.writeNumberField("Resource Profile Id", executorInfo.resourceProfileId)