… mllib/src/test/scala/org/apache/spark/ml/MLEventsSuite.scala
… sql/core/src/test/scala/org/apache/spark/sql/execution/SQLJsonProtocolSuite.scala
… sql/core/src/test/scala/org/apache/spark/sql/execution/metric/SQLMetricsSuite.scala
… sql/core/src/test/scala/org/apache/spark/sql/execution/ui/SQLAppStatusListenerSuite.scala
… sql/core/src/test/scala/org/apache/spark/sql/streaming/StreamingQueryListenerSuite.scala