@JoshRosen 2022-06-30
I submitted https://github.com/apache/spark/pull/37027 to fix the [pre-existing JsonProtocol bugs that I found during my testing](https://github.com/apache/spark/pull/36885#issuecomment-1167855113). Once that PR merges I will fix conflicts in this PR.