@JoshRosen 2022-06-30
I've fixed merge conflicts and pushed another change to add typechecking during JSON node extraction: this doesn't matter when parsing well-formed / expected JSON, but could make a difference if invalid JSON is presented; see https://github.com/apache/spark/pull/36885#discussion_r911493733 for details.

I removed the "WIP" tag; I think this PR should now be ready for a final round of review.