@JoshRosen 2022-07-03
Thanks everyone. I'm merging this to master (Spark 3.4.0).

(The latest test failure is unrelated / a pre-existing flaky test; my most recent commit only added a code comment)