@dongjoon-hyun 2023-02-11
SPARK-42403 is filed as a release blocker for Apache Spark 3.4.0.