@JoshRosen 2023-02-11
@dongjoon-hyun, thanks for the ping. I think that I spot the problem: it looks like the `File Name` field could contain a `null` value and the old code was likely handling JNothing -> `null.asInstanceOf[String]` implicitly.

I think we can fix this by either modifying `extractString` to handle this case or by special-casing this for `FileName`. I'd prefer the former approach because it's probably more robust.