@dongjoon-hyun 2023-02-11
Ya, actaully, I tried to relax the `requirement` in `extractString`, but it could miss other cases.
So, I'm currently proposing this one.
- https://github.com/apache/spark/pull/39972