@dongjoon-hyun 2023-02-11
I'll convert it to a normal PR after adding a test case properly.