@JoshRosen 2023-02-11
I like your approach as a surgical fix.

I'm trying to think through if there are any other cases where we can get `null` values for strings:

- It can't happen for `Properties` because Java `Properties` forbids `null` values.
- Many of the listener event string fields are things like executorId, block IDs, hostnames, etc., none of which can possibly be null at the time that the event logs are written.
- We _might_ get nulls in the `ExceptionFailure` TaskEndReason because an exception can have a `null` message.

Given this, I'm thinking that I'd prefer to fix this by changing `extractString` to check whether the value is textual or null.