@JoshRosen 2023-02-11
I've submitted my proposed fix at https://github.com/apache/spark/pull/39973